Add native meson.build, draft PR
This is a preview 'sanity check' for review by meson devs
(I'm a build and CMake novice with intermediate meson).

This draft PR adds native meson build support for clients
(not for snitch development; no test targets are defined).

Meson build is already supported via meson's cmake module:

    if get_option('snitch_cmake')
      snitch_cmake = import('cmake').subproject('snitch')
      snitch_dep = snitch_cmake.dependency('snitch')
    else
      snitch_dep = dependency('snitch')
    endif

(I see no way to put this client logic inside snitch.)

This PR merely adds a way for clients to build directly with
meson build and without any CMake dependency installed.

Note that doctest has seven lines of meson support:

    project('doctest', ['cpp'], version: '2.4.9')
    doctest_dep = declare_dependency(
      include_directories: include_directories('doctest'))

    if meson.version().version_compare('>=0.54.0')
      meson.override_dependency('doctest', doctest_dep)
    endif

doctest also has an update script to edit the version number.

A neater way to automate version number updates for snitch
might be to place the version number in a file, to be read
by both CMake and meson setup. Otherwise, the version has to
be updated in both CMakeLists.txt and meson.build files -
a bit of maintenance burden easily forgotten.

The main bloat in this PR is in support of build options
(it's likely I've missed a more direct way).

This PR copies the configuration options of the CMake build,
replicating configure_file data as meson_options variables.
This calls for repetitive code to copy to and from variables;
a maintenance burden when options are added or if the default
values are changed, as they need to be edited in many places.

In order to generate snitch_config.hpp in build/snitch dir
a snitch subdirectory in the source root had to be added
(at least I couldn't find a way to configure_file elsewhere).
This, and the meson_options.txt file, all add 'noise' to the
snitch source root, something I find can be offputting.

* Is the utility of the options worth the cost?
* Can some of the options be discarded for meson support?

On the other hand, support for generating the snitch_all.hpp
single file header is straightforward using a custom_target
to invoke meson's Python module to run make_snitch_all.py

I'll post a PR link on meson matrix channel once this is up.
The root meson.build is still longer than ideal at 64 lines;
is there any redundancy 'cargo-culted' in?
Or, is there anything missing that should be added.

Thanks for reading (this little essay is longer than ideal)!
Please contribute your thoughts in comments.