Fix install of snitch_all.hpp
The snitch_all.hpp custom_target was not being installed;
the option install:true was needed. as well as install_dir:.
However, the install now failed with 'cannot find snitch.hpp'.
The solution was to move the custom_target to ./snitch/meson.build
(decluttering the main script too).

This commit also removes single-use 'python' and 'pckconfig' vars
and adds multiple-use include_dirs and main vars to aid readability.
A make_snitch_all file variable is also defined for explicitness.

The duplicated description strings on conf_data vars are deleted,
because they're not used when configuring with a template file,
allowing configuration_data to be set from a dictionary of pairs.

snitch_single_header_dep is removed; I doubt anyone will use it.