minimal emulation of Catch2 test listing
This is the minimal emulation of the list-tests command line option that
CLion needs to discover and list tests. CLion calls the test binary with
the command line `--list-tests --order lex ~[.]` before running tests.
It doesn't seem to mind that the tests aren't actually ordered, but it
does break it's parsing if:
  - there is no leading "Matching test cases:"
  - tags aren't listed with an indent
  - tests are listed without tags