remove fallback for floating point std::to_char
If the compiler doesn't support std::to_char forfloating point, then the
compile time option should disable use of std::to_char for all conversions.
This avoids mixing and matching implementations. Either append_constexpr
is used for all conversion, or std::to_char is used for all conversions.

This slightly complicates the testing as there is now a two variable truth
table as to the expected results when formatting -0.0.