Fix ambiguous overloads on 32-bit platforms
On the 32-bit platforms in CI (windows and emscripten) large_int_t and
large_uint_t seem to alias each other in function definitions. Fix this by
using signed/unsigned concepts for the parameter. That way the incorrect
function will not be considered.

This required small bug fix in snitch::concepts. signed_integral and
unsigned_integral concepts need to check for sign and also make sure the
type is integral so that floating point or other types are not considered.