Add static instance of console reporter to registry
This works around the need for there to be an instantiation of the
console reporter now that it has state. Since the console reporter is
the default reporter it can be statically initialized in the registry.

To allow the registry to continue being constinit, a defaulted constructor
was added to the console reporter. There are also static functions in the
registry that are used to register the console reporter as a reporter
with the existing API.