non static console reporter
The reporter needed to be static to allow the assignment of it's functions
to the function references. But the function reference type actually does
have storage for a single pointer that could be used to store the `this`
pointer of the registery. To do this we need to be sure and invoke the
appropriate constructor for the function reference.